requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
boto3>=1.26.0
s3fs>=2023.0.0
pyarrow>=13.0.0
joblib>=1.3.0
mlflow>=2.11.0
xgboost>=1.7
hydra-core>=1.3.2
shap>=0.41.0
matplotlib>=3.5.0